requirements.txt
fastapi
uvicorn
python-jose[cryptography]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-dotenv
alembic